remove signature scan
The entire point of signature scanning is removing any and all use of pointers, essentially ensuring future proofing. Using signatures *and* pointers (like you are doing) is pointless and you are monopolizing CPU power for nothing. The pointer to the current room you found is `g_RuntimeSceneManager.GetActiveScene.name`. The "load state" you were scanning for is `g_RuntimeSceneManager.sceneCount` (the amount of currently loaded scenes, which will be `> 1` whenever it's loading). As you can see, no need to scan for something that you already have a pointer to.